requirements.txt
feedparser
google-generativeai
sqlite-utils
slack_sdk
requests
python-dotenv
